backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
motor>=3.3.2
pymongo>=4.6.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.5.0
opencv-python>=4.8.1
numpy>=2.0.0
rapidfuzz>=3.5.2
pytesseract>=0.3.10
pillow>=10.1.0
email-validator>=2.0.0
# Note: mediapipe and paddleocr may not be available for Python 3.14+
# These are optional dependencies for advanced ML features
# mediapipe>=0.10.8
# paddleocr>=2.7.3.3

